+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - enforce signed and unsigned integers in MultiWavecalCalibration (resolve XL
       config files with floating-point horizontal ROI)
     - update BatchCollection and Scan Averaging for AndorDevice
+    - added EEPROM.laser_interlock_excluded
 - 2025-04-08 2.3.0
     - IDS cameras
         - add IDSDevice, IDSCamera